% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.khazardcond.R
\name{plot.khazardcond}
\alias{plot.khazardcond}
\title{Plot of kernel conditional hazard estimate from an object of class khazardcond}
\usage{
\method{plot}{khazardcond}(x, type = "persp", zlim, xlab, ylab, zlab,
  main, ...)
}
\arguments{
\item{x}{Object of class khazardcond}

\item{type}{type of plot. Possible types are:  "persp" (default), "persp3d", "contour".}

\item{zlim}{Limits for the z axis.}

\item{xlab}{Label for the x axis.}

\item{ylab}{Label for the y axis.}

\item{zlab}{Label for the z axis.}

\item{main}{Title of plot.}

\item{...}{Additional arguments.}
}
\description{
Plot of kernel conditional hazard estimate from an object of class khazardcond
}
\examples{
library(survival)
fit<-khazardcond(times = lung$time,delta = lung$status-1,covariate = lung$age,h=c(200,20))
plot(fit)
}
\seealso{
\code{\link{khazardcond}}
}
